# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/pulse/ingestor": {
            "post": {
                "description": "Ingestor de pulsos",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Pulso"
                ],
                "summary": "Ingestor de pulsos",
                "parameters": [
                    {
                        "description": "Pulse",
                        "name": "pulse",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_pulse.Pulse"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    }
                }
            }
        }
    },
    "definitions": {
        "internal_pulse.Pulse": {
            "type": "object",
            "required": [
                "product_sku",
                "tenant_id",
                "use_unit",
                "used_amount"
            ],
            "properties": {
                "product_sku": {
                    "description": "ProductSku é o SKU do produto, geralmente segue o padrão \"SKU-\u003cnumero\u003e\"",
                    "type": "string"
                },
                "tenant_id": {
                    "type": "string"
                },
                "use_unit": {
                    "description": "UseUnit é a unidade utilizada para o valor utilizado do produto",
                    "allOf": [
                        {
                            "$ref": "#/definitions/internal_pulse.PulseUnit"
                        }
                    ]
                },
                "used_amount": {
                    "description": "UsedAmount é o valor utilizado do produto",
                    "type": "number"
                }
            }
        },
        "internal_pulse.PulseUnit": {
            "type": "string",
            "enum": [
                "KB",
                "MB",
                "GB",
                "KB/sec",
                "MB/sec",
                "GB/sec"
            ],
            "x-enum-varnames": [
                "KB",
                "MB",
                "GB",
                "KBxSec",
                "MBxSec",
                "GBxSec"
            ]
        }
    }
}